cleaning code, restructuring projects
- split libAudio into separate VS project (generates .lib file)
- moved AUDDRV defines into Makefile/projects
- moved VS object files into obj/ directory, final libraries to bin/
- added 64-bit build to VS2010 project

+++ audio/AudDrv_WASAPI.cpp
@@ -138,7 +138,7 @@ UINT8 WASAPI_Init(void)
 	UINT devCount;
 	UINT32 curDev;
 	UINT32 devLstID;
-	UINT32 devNameSize;
+	size_t devNameSize;
 	IMMDeviceEnumerator* devEnum;
 	IMMDeviceCollection* devList;
 	IMMDevice* audDev;
